+++ MechanicWebAppAPI/Core/Dtos/Repair/RepairDto.cs
@@ -6,7 +6,7 @@ namespace MechanicWebAppAPI.Core.Dtos.Repair
     {
         public string Repair_description { get; set; }
 
-        public int Repair_cost { get; set; }
+        public decimal Repair_cost { get; set; }
 
         public string Repair_date { get; set; }
